app/requirements.txt
flask == 3.0.3
python-dotenv == 1.0.1
uwsgi == 2.0.25.1